multiple updates
1 new interfaces
2 fixed feedback logic and ui  fixed interview logic

+++ src/main/java/Models/Interview.java
@@ -4,9 +4,9 @@ import java.time.LocalDateTime;
 
 public class Interview {
 
-    private int id;
-    private int applicationId;
-    private int recruiterId;
+    private Long id;
+    private Long applicationId;
+    private Long recruiterId;
     private LocalDateTime scheduledAt;
     private int durationMinutes;
     private String mode;
@@ -17,7 +17,7 @@ public class Interview {
 
     public Interview() {}
 
-    public Interview(int applicationId, int recruiterId,
+    public Interview(Long applicationId, Long recruiterId,
                      LocalDateTime scheduledAt, int durationMinutes, String mode) {
         this.applicationId = applicationId;
         this.recruiterId = recruiterId;
@@ -28,14 +28,14 @@ public class Interview {
     }
 
     // Getters & Setters
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
-    public int getApplicationId() { return applicationId; }
-    public void setApplicationId(int applicationId) { this.applicationId = applicationId; }
+    public Long getApplicationId() { return applicationId; }
+    public void setApplicationId(Long applicationId) { this.applicationId = applicationId; }
 
-    public int getRecruiterId() { return recruiterId; }
-    public void setRecruiterId(int recruiterId) { this.recruiterId = recruiterId; }
+    public Long getRecruiterId() { return recruiterId; }
+    public void setRecruiterId(Long recruiterId) { this.recruiterId = recruiterId; }
 
     public LocalDateTime getScheduledAt() { return scheduledAt; }
     public void setScheduledAt(LocalDateTime scheduledAt) { this.scheduledAt = scheduledAt; }
